+++ src/actions/product.ts
@@ -1,5 +1,8 @@
+"use server";
+
 import { Product } from "@/types/types";
-import { instance } from "./Client";
+import { instance } from "../services/Client";
+import { serverInstance } from "@/lib/axios";
 
 interface ProductQuery {
   categoryId?: string[];
@@ -8,17 +11,17 @@ interface ProductQuery {
 }
 
 export const getProducts = async (params: ProductQuery): Promise<Product[]> => {
-  const response = await instance.get(`/catalogue/client/products`, {
+  const response = await serverInstance.get(`/catalogue/client/products`, {
     params: params,
     paramsSerializer: { indexes: false },
   });
   return response.data.data;
 };
 export const getProduct = async (productId: string): Promise<Product> => {
-  const response = await instance.get(`/catalogue/client/products/${productId}`);
+  const response = await serverInstance.get(`/catalogue/client/products/${productId}`);
   return response.data.data;
 };
 export const getProductBySKU = async (productSKU: string): Promise<Product> => {
-  const response = await instance.get(`/catalogue/client/products/sku/${productSKU}`);
+  const response = await serverInstance.get(`/catalogue/client/products/sku/${productSKU}`);
   return response.data.data;
 };
