+++ src/app/(route)/order-summary/[orderSummaryId]/components/payment-options.tsx
@@ -5,8 +5,9 @@ import { ArrowRight } from "lucide-react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { toast } from "react-hot-toast";
 import useCart from "@/hooks/use-cart";
-import { createOrder } from "@/services/PaymentService";
+import { createOrder } from "@/actions/payment";
 import { AxiosError } from "axios";
+import { Button, Spinner } from "flowbite-react";
 
 function PaymentOptions({ orderSummaryId }: { orderSummaryId: string }) {
   const [paymentGateway, setPaymentGateway] = useState<string>("");
@@ -28,7 +29,6 @@ function PaymentOptions({ orderSummaryId }: { orderSummaryId: string }) {
     }
   }, [searchParams, removeAll]);
 
-
   const onProceed = async () => {
     try {
       setIsLoading(true);
@@ -37,12 +37,13 @@ function PaymentOptions({ orderSummaryId }: { orderSummaryId: string }) {
         orderSummaryId: orderSummaryId,
         phone: "51234567",
         address: "hello",
-        paymentGateway: "TAP",
+        paymentGateway,
       };
       const response = await createOrder(payload);
       window.location = response?.paymentUrl;
     } catch (e) {
       setIsLoading(false);
+      console.error(e);
       const error = e as AxiosError;
       const msg: any = error?.response?.data;
       const errorMessage = msg?.message || "";
@@ -56,12 +57,12 @@ function PaymentOptions({ orderSummaryId }: { orderSummaryId: string }) {
     <div className="w-full  mt-8 md:mt-0 pl-4">
       <h2 className="font-semibold text-xl">Payment Method</h2>
       <ul className="grid w-full gap-6 my-4">
-        <li>
+        <li onClick={() => setPaymentGateway("TAP")}>
           <input
             type="radio"
-            checked={paymentGateway === "TAP" ? true : false}
+            checked={paymentGateway == "TAP" ? true : false}
             className="hidden peer"
-            onClick={() => setPaymentGateway("TAP")}
+            // onClick={() => setPaymentGateway("TAP")}
           />
           <label className="inline-flex items-center justify-between w-full p-5 text-gray-500 bg-white border border-gray-200 rounded-lg cursor-pointer dark:hover:text-gray-300 dark:border-gray-700 dark:peer-checked:text-blue-500 peer-checked:border-blue-600 peer-checked:text-blue-600 hover:text-gray-600 hover:bg-gray-100 dark:text-gray-400 dark:bg-gray-800 dark:hover:bg-gray-700">
             <div className="block">
@@ -84,14 +85,20 @@ function PaymentOptions({ orderSummaryId }: { orderSummaryId: string }) {
       </ul>
 
       <div className="mt-8">
-        <button
-          disabled={isLoading}
+        <Button
           onClick={onProceed}
-          type="button"
-          className="w-full bg-blue-500 text-white font-semibold py-2 rounded-md hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-400"
+          disabled={paymentGateway.length === 0 || isLoading}
+          className="w-full mt-6"
         >
-          Place Order
-        </button>
+          {isLoading ? (
+            <>
+              <Spinner aria-label="signing" />
+              <span className="pl-3">Please wait...</span>
+            </>
+          ) : (
+            " Place Order"
+          )}
+        </Button>
       </div>
     </div>
   );
