+++ src/app/(route)/profile/page.tsx
@@ -1,4 +1,5 @@
-import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+
+import { authOptions } from "@/app/api/auth/[...nextauth]/auth-options"
 import { getServerSession } from "next-auth"
